undo the changings
