+++ bindings/python/core_types/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(LIB_NAME py_core_types)
 
 set(SRC_Headers
-    include/py_core_types/py_queued_event_handler.h
     include/py_core_types/py_core_types.h
     #
     include/py_core_types/py_base_object.h
@@ -10,6 +9,8 @@ set(SRC_Headers
     include/py_core_types/py_list_object.h
     include/py_core_types/py_procedure.h
     include/py_core_types/py_function.h
+    #
+    include/py_core_types/py_queued_event_handler.h
     include/py_core_types/py_queued_event_handler_impl.h
     include/py_core_types/py_event_queue.h
 )
