+++ bindings/python/core_types/include/py_core_types/py_event_handler_impl.h
@@ -39,9 +39,8 @@ public:
     {
         try
         {
-            if(sender) sender->addRef();
-            if(eventArgs) eventArgs->addRef();
-
+            ObjectPtr<IBaseObject> senderPtr(sender);
+            ObjectPtr<IEventArgs> eventArgsPtr(eventArgs);
             subscription(sender, eventArgs);
         }
         catch (const DaqException& e)
