--- bindings/python/opendaq/include/py_context/py_context.h
@@ -1,30 +1,0 @@
-/*
- * Copyright 2022-2025 openDAQ d.o.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-#include <opendaq/context.h>
-#include "py_core_types/py_queued_event_handler.h"
-
-BEGIN_NAMESPACE_OPENDAQ
-
-// workaround to have a separate interface for the queued event handler
-DECLARE_OPENDAQ_INTERFACE(IPythonContext, IContext)
-{
-    virtual ErrCode INTERFACE_FUNC setEventToQueue(IPythonQueuedEventHandler* eventHandler, IBaseObject* sender, IEventArgs* eventArgs) = 0;
-    virtual ErrCode INTERFACE_FUNC processEventsFromQueue() = 0;
-};
-
-END_NAMESPACE_OPENDAQ
