--- bindings/python/opendaq/include/py_opendaq/py_instance_builder_impl.h
@@ -1,32 +1,0 @@
-/*
- * Copyright 2022-2025 openDAQ d.o.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include <opendaq/instance_builder_impl.h>
-
-BEGIN_NAMESPACE_OPENDAQ
-
-class PythonInstanceBuilderImpl: public InstanceBuilderImpl
-{
-public:
-    using Super = InstanceBuilderImpl;
-    using Super::Super;
-
-    virtual ErrCode INTERFACE_FUNC build(IInstance** instance) override;
-};
-
-END_NAMESPACE_OPENDAQ
