+++ bindings/python/opendaq/py_opendaq.cpp
@@ -40,7 +40,6 @@ void wrapDaqComponentOpenDaq(pybind11::module_ m)
     auto classIConfigProvider = declareIConfigProvider(m);
     auto classIInstance = declareIInstance(m);
     auto classIContext = declareIContext(m);
-    auto classIPythonContext = declareIPythonContext(m);
     auto classIDeviceDomain = declareIDeviceDomain(m);
     auto classIDeviceInfo = declareIDeviceInfo(m);
     auto classIDeviceInfoConfig = declareIDeviceInfoConfig(m);
@@ -144,7 +143,6 @@ void wrapDaqComponentOpenDaq(pybind11::module_ m)
     defineIInstance(m, classIInstance);
     defineIConfigProvider(m, classIConfigProvider);
     defineIContext(m, classIContext);
-    defineIPythonContext(m, classIPythonContext);
     defineIDeviceDomain(m, classIDeviceDomain);
     defineIDeviceInfo(m, classIDeviceInfo);
     defineIDeviceInfoConfig(m, classIDeviceInfoConfig);
